refactor(merk): split versioned avg-case cost fns into per-version files (platform-style)
Restructures `merk/src/estimated_costs/average_case_costs.rs` from a
single file with inline version branches / co-located v* functions
into the dash-platform-style layout: one directory per versioned
function with a `mod.rs` dispatcher and per-version `v0.rs`/`v1.rs`/
`v2.rs` files.

```
merk/src/estimated_costs/average_case_costs/
├── mod.rs                                          # types, non-versioned methods, helpers, tests
├── estimated_sum_trees_size/                       # EstimatedSumTrees::estimated_size
│   ├── mod.rs                                      # version dispatcher
│   ├── v0.rs                                       # grove v1 formula
│   ├── v1.rs                                       # grove v2 formula
│   └── v2.rs                                       # grove v3 formula (adds Provable* weights)
├── value_with_feature_and_flags_size/              # EstimatedLayerSizes::value_with_...
│   ├── mod.rs
│   ├── v0.rs                                       # legacy unweighted Mix average
│   └── v1.rs                                       # proper weighted Mix average
└── add_average_case_merk_propagate/                # free fn dispatcher
    ├── mod.rs                                      # also exports average_case_merk_propagate wrapper
    ├── v0.rs                                       # grove v1
    ├── v1.rs                                       # grove v2
    └── v2.rs                                       # grove v3 (Mix divisor fix)
```

`EstimatedSumTrees::estimated_size` was the one CodeRabbit-style
target: previously it had an inline `if version == 0 { ... } else if
version == 1 { ... }` chain after a `check_grovedb_v0_v1_or_v2!`
macro. It now follows the same dispatcher-via-match pattern as the
other two functions (and the per-version logic is split across three
files for readability).

No behavior change — code was moved as-is. All 657 merk tests + 1834
grovedb tests pass.
